Reflect refactoring throughout codebase
