+++ src/main/java/com/google/sps/ContextListener.java
@@ -5,7 +5,7 @@ import com.google.sps.model.queue.EntityDbQueue;
 import com.google.sps.model.queue.WantToWatchQueueObject;
 import com.google.sps.model.queue.WatchedQueueObject;
 import com.google.sps.model.review.ReviewObject;
-import com.google.sps.servlets.user.UserObject;
+import com.google.sps.model.user.UserObject;
 import com.googlecode.objectify.ObjectifyService;
 
 import javax.servlet.ServletContextEvent;
